Control post-preview cursor positioning
I'm a heavy user of
[pip-mysql.vim](https://github.com/NLKNguyen/pipe-mysql.vim) and one
thing that's been bugging me is that after running a query, the cursor
is at the bottom of the buffer.  This makes sense if I use this plugin
to run terminal commands and want to see the final output but in the
case of mysql queries, it's much more useful to be at the top where I
can see the columns returned.

This commit allows the consumer of `Pipe` to define if they want the
cursor to be at the top or bottom.  This can be set either via a global
variable `g:top_or_bottom`, or by passing a second parameter to
`g:Pipe`. Default functionality is unchanged from the current so there
should be no breaking changes from this commit.